Prediction_and_Localisation/Kafka/requirements.txt
aiokafka==0.8.1
fastapi==0.104.1
kafka_python==2.0.2
redis==5.0.1
websockets==12.0
uvicorn